apps/backend/pyproject.toml
[project]
name = "rm-backend"
version = "1.2.0"
description = "Resume Matcher Backend API"
requires-python = ">=3.13"
dependencies = [
    "fastapi==0.128.4",
    "uvicorn==0.40.0",
    "python-multipart==0.0.26",
    "pydantic==2.12.5",
    "pydantic-settings==2.12.0",
    "firebase-admin>=6.0.0",
    "litellm==1.83.0",
    "markitdown[docx]==0.1.4",
    "pdfminer.six==20260107",
    "playwright==1.58.0",
    "python-docx==1.2.0",
    "python-dotenv==1.2.1",
    "firebase-admin>=7.4.0",
    "passlib>=1.7.4",
    "bcrypt>=5.0.0",
    "python-jose>=3.5.0",
    "email-validator>=2.3.0",
]

[project.scripts]
app = "app.main:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = [
    "--strict-markers",
    "-v",
]
markers = [
    "unit: pure function tests, no external dependencies",
    "service: service layer tests with mocked LLM",
    "integration: API endpoint tests with httpx AsyncClient",
]
